requirements.txt
pygame-ce>=2.4.0
sympy>=1.12
numpy>=1.26.3